Do not run buildbot as root
All compilation stuff could be done as buildbot user.
Only docker actions need spetial handling.
They do not need to be root, but running root permits to read docker socket in /run.
So we need to allow buildbot user to read it by adding to it the docker group.
BUT docker group of host and container could have a different GID.
So we need to sync GID of container to what host uses.

We cannot do anything at runtime (too complex to hack buildbot entrypoint)
so we need to did this "sync" at build time via a docker ARG.

Closes: #5